+++ src/main/java/nextstep/laddergame/domain/LadderGame.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame;
+package nextstep.laddergame.domain;
 
 public record LadderGame(Participants participants, Ladder ladder) {
