+++ src/main/java/nextstep/laddergame/domain/Participant.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame;
+package nextstep.laddergame.domain;
 
 public record Participant(String name) {
