+++ src/main/java/nextstep/laddergame/domain/Participants.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame;
+package nextstep.laddergame.domain;
 
 import java.util.Arrays;
 import java.util.List;
