+++ src/test/java/nextstep/laddergame/LadderTest.java
@@ -1,6 +1,7 @@
 package nextstep.laddergame;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import nextstep.laddergame.domain.Ladder;
 import org.junit.jupiter.api.Test;
 
 class LadderTest {
