+++ src/test/java/nextstep/laddergame/LineTest.java
@@ -3,6 +3,7 @@ package nextstep.laddergame;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
+import nextstep.laddergame.domain.Line;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
