+++ src/test/java/nextstep/laddergame/ParticipantTest.java
@@ -1,6 +1,7 @@
 package nextstep.laddergame;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import nextstep.laddergame.domain.Participant;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
