+++ src/test/java/nextstep/laddergame/ParticipantsTest.java
@@ -1,6 +1,7 @@
 package nextstep.laddergame;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import nextstep.laddergame.domain.Participants;
 import org.junit.jupiter.api.Test;
 
 class ParticipantsTest {
